disk image may well extend beyond 4 GiB

+++ disk/shfs/shfs.c
@@ -66,7 +66,8 @@ unsigned int fsStart;		/* file system start sector */
 void readBlock(FILE *disk,
                EOS32_daddr_t blockNum,
                unsigned char *blockBuffer) {
-  fseek(disk, fsStart * SECTOR_SIZE + blockNum * BLOCK_SIZE, SEEK_SET);
+  fseek(disk, (unsigned long) fsStart * SECTOR_SIZE +
+        (unsigned long) blockNum * BLOCK_SIZE, SEEK_SET);
   if (fread(blockBuffer, BLOCK_SIZE, 1, disk) != 1) {
     error("cannot read block %lu (0x%lX)", blockNum, blockNum);
   }
